GVT-2411 Big honkin' switch linking refactor
