+++ infra/src/main/kotlin/fi/fta/geoviite/infra/tracklayout/LocationTrackService.kt
@@ -10,6 +10,7 @@ import fi.fta.geoviite.infra.linking.LocationTrackEndpoint
 import fi.fta.geoviite.infra.linking.LocationTrackPointUpdateType.END_POINT
 import fi.fta.geoviite.infra.linking.LocationTrackPointUpdateType.START_POINT
 import fi.fta.geoviite.infra.linking.LocationTrackSaveRequest
+import fi.fta.geoviite.infra.linking.TopologyLinkFindingSwitch
 import fi.fta.geoviite.infra.logging.serviceCall
 import fi.fta.geoviite.infra.math.*
 import fi.fta.geoviite.infra.publication.ValidationVersion
@@ -478,17 +479,16 @@ class LocationTrackService(
         alignment: LayoutAlignment,
         startChanged: Boolean = false,
         endChanged: Boolean = false,
-        overlaidTracks: Map<IntId<LocationTrack>, Pair<LocationTrack, LayoutAlignment>> = mapOf(),
     ): LocationTrack {
         val nearbyTracksAroundStart =
             (alignment.start?.let(::fetchNearbyLocationTracksWithAlignments)?.filter { (nearbyLocationTrack, _) ->
                 nearbyLocationTrack.id != track.id
-            } ?: listOf()).map { nearbyTrack -> overlaidTracks.getOrDefault(nearbyTrack.first.id, nearbyTrack)  }
+            } ?: listOf())
 
         val nearbyTracksAroundEnd =
             (alignment.end?.let(::fetchNearbyLocationTracksWithAlignments)?.filter { (nearbyLocationTrack, _) ->
                 nearbyLocationTrack.id != track.id
-            } ?: listOf()).map { nearbyTrack -> overlaidTracks.getOrDefault(nearbyTrack.first.id, nearbyTrack)  }
+            } ?: listOf())
 
         return calculateLocationTrackTopology(
             track, alignment, startChanged, endChanged, NearbyTracks(
@@ -600,18 +600,23 @@ fun calculateLocationTrackTopology(
     startChanged: Boolean = false,
     endChanged: Boolean = false,
     nearbyTracks: NearbyTracks,
+    newSwitch: TopologyLinkFindingSwitch? = null,
 ): LocationTrack {
     val startPoint = alignment.firstSegmentStart
     val endPoint = alignment.lastSegmentEnd
     val ownSwitches = alignment.segments.mapNotNull { segment -> segment.switchId }.toSet()
 
     val startSwitch = if (!track.exists || startPoint == null) null
-    else if (startChanged) findBestTopologySwitchMatch(startPoint, ownSwitches, nearbyTracks.aroundStart, null)
-    else findBestTopologySwitchMatch(startPoint, ownSwitches, nearbyTracks.aroundStart, track.topologyStartSwitch)
+    else if (startChanged) {
+        findBestTopologySwitchMatch(startPoint, ownSwitches, nearbyTracks.aroundStart, null, newSwitch)
+    }
+    else findBestTopologySwitchMatch(startPoint, ownSwitches, nearbyTracks.aroundStart, track.topologyStartSwitch, newSwitch)
 
     val endSwitch = if (!track.exists || endPoint == null) null
-    else if (endChanged) findBestTopologySwitchMatch(endPoint, ownSwitches, nearbyTracks.aroundEnd, null)
-    else findBestTopologySwitchMatch(endPoint, ownSwitches, nearbyTracks.aroundEnd, track.topologyEndSwitch)
+    else if (endChanged) {
+        findBestTopologySwitchMatch(endPoint, ownSwitches, nearbyTracks.aroundEnd, null, newSwitch)
+    }
+    else findBestTopologySwitchMatch(endPoint, ownSwitches, nearbyTracks.aroundEnd, track.topologyEndSwitch, newSwitch)
 
     return if (track.topologyStartSwitch == startSwitch && track.topologyEndSwitch == endSwitch) {
         track
@@ -629,10 +634,11 @@ fun findBestTopologySwitchMatch(
     ownSwitches: Set<DomainId<TrackLayoutSwitch>>,
     nearbyTracksForSearch: List<Pair<LocationTrack, LayoutAlignment>>,
     currentTopologySwitch: TopologyLocationTrackSwitch?,
+    newSwitch: TopologyLinkFindingSwitch?,
 ): TopologyLocationTrackSwitch? {
     val defaultSwitch = if (currentTopologySwitch?.switchId?.let(ownSwitches::contains) != false) null
     else currentTopologySwitch
-    return findBestTopologySwitchFromSegments(target, ownSwitches, nearbyTracksForSearch) ?: defaultSwitch
+    return findBestTopologySwitchFromSegments(target, ownSwitches, nearbyTracksForSearch, newSwitch) ?: defaultSwitch
     ?: findBestTopologySwitchFromOtherTopology(target, ownSwitches, nearbyTracksForSearch)
 }
 
@@ -640,9 +646,10 @@ private fun findBestTopologySwitchFromSegments(
     target: IPoint,
     ownSwitches: Set<DomainId<TrackLayoutSwitch>>,
     nearbyTracks: List<Pair<LocationTrack, LayoutAlignment>>,
+    newSwitch: TopologyLinkFindingSwitch?,
 ): TopologyLocationTrackSwitch? = nearbyTracks.flatMap { (_, otherAlignment) ->
     otherAlignment.segments.flatMap { segment ->
-        if (segment.switchId !is IntId || ownSwitches.contains(segment.switchId)) listOf()
+        if (segment.switchId !is IntId || ownSwitches.contains(segment.switchId) || segment.switchId == newSwitch?.id) listOf()
         else listOfNotNull(
             segment.startJointNumber?.let { number ->
                 pickIfClose(
@@ -651,7 +658,9 @@ private fun findBestTopologySwitchFromSegments(
             },
             segment.endJointNumber?.let { number -> pickIfClose(segment.switchId, number, target, segment.segmentEnd) },
         )
-    }
+    } + (newSwitch?.joints?.mapNotNull { sj ->
+        pickIfClose(newSwitch.id, sj.number, target, sj.location)
+    } ?: listOf())
 }.minByOrNull { (_, distance) -> distance }?.first
 
 private fun findBestTopologySwitchFromOtherTopology(
