+++ ui/src/tool-panel/switch/geometry-switch-linking-infobox.tsx
@@ -5,12 +5,7 @@ import { useTranslation } from 'react-i18next';
 import styles from './switch-infobox.scss';
 import InfoboxButtons from 'tool-panel/infobox/infobox-buttons';
 import { Button, ButtonSize, ButtonVariant } from 'vayla-design-lib/button/button';
-import {
-    LinkingState,
-    LinkingSwitch,
-    SuggestedSwitch,
-    SwitchLinkingSegment,
-} from 'linking/linking-model';
+import { LinkingState, LinkingSwitch, SuggestedSwitch } from 'linking/linking-model';
 import { IconColor, Icons } from 'vayla-design-lib/icon/Icon';
 import { SwitchEditDialogContainer } from './dialog/switch-edit-dialog';
 import { LayoutSwitch } from 'track-layout/track-layout-model';
@@ -106,6 +101,7 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
         }
         return undefined;
     }, [geometrySwitchId, selectedSuggestedSwitch, geometrySwitchLayout]);
+    const switchStructure = useSwitchStructure(suggestedSwitch?.switchStructureId);
     const [showAddSwitchDialog, setShowAddSwitchDialog] = React.useState(false);
     const [linkingCallInProgress, setLinkingCallInProgress] = React.useState(false);
     const selectedLayoutSwitch = useSwitch(linkingState?.layoutSwitchId, publishType);
@@ -114,15 +110,17 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
     );
     const switchTypeMatch =
         suggestedSwitch &&
+        switchStructure &&
         selectedLayoutSwitch &&
-        suggestedSwitch.switchStructure.id == selectedLayoutSwitch.switchStructureId
+        switchStructure.id == selectedLayoutSwitch.switchStructureId
             ? SwitchTypeMatch.Exact
             : suggestedSwitch &&
-              selectedLayoutSwitchStructure &&
-              suggestedSwitch.switchStructure.baseType == selectedLayoutSwitchStructure.baseType &&
-              suggestedSwitch.switchStructure.hand == selectedLayoutSwitchStructure.hand
-            ? SwitchTypeMatch.Similar
-            : SwitchTypeMatch.Invalid;
+                selectedLayoutSwitchStructure &&
+                switchStructure &&
+                switchStructure.baseType == selectedLayoutSwitchStructure.baseType &&
+                switchStructure.hand == selectedLayoutSwitchStructure.hand
+              ? SwitchTypeMatch.Similar
+              : SwitchTypeMatch.Invalid;
 
     const [switchTypeDifferenceIsConfirmed, setSwitchTypeDifferenceIsConfirmed] =
         React.useState(false);
@@ -151,23 +149,9 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
 
     async function link() {
         if (linkingState?.suggestedSwitch && linkingState.layoutSwitchId) {
-            const params = {
-                switchStructureId: linkingState.suggestedSwitch.switchStructure.id,
-                geometryPlanId: planId,
-                geometrySwitchId: linkingState.suggestedSwitch.geometrySwitchId,
-                layoutSwitchId: linkingState.layoutSwitchId,
-                joints: linkingState.suggestedSwitch.joints.map((joint) => {
-                    return {
-                        jointNumber: joint.number,
-                        location: joint.location,
-                        segments: joint.matches satisfies SwitchLinkingSegment[],
-                        locationAccuracy: joint.locationAccuracy,
-                    };
-                }),
-            };
             setLinkingCallInProgress(true);
             try {
-                await linkSwitch(params);
+                await linkSwitch(linkingState.suggestedSwitch, linkingState.layoutSwitchId);
                 SnackBar.success('tool-panel.switch.geometry.linking-succeed-msg');
                 onStopLinking();
             } finally {
@@ -178,20 +162,23 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
 
     return (
         <React.Fragment>
-            {selectedSuggestedSwitch?.alignmentEndPoint && onSuggestedSwitchChange && (
-                <GeometrySwitchLinkingSuggestedInfobox
-                    contentVisible={visibilities.suggestedSwitch}
-                    onContentVisibilityChange={() =>
-                        onVisibilityChange({
-                            ...visibilities,
-                            suggestedSwitch: !visibilities.suggestedSwitch,
-                        })
-                    }
-                    suggestedSwitch={selectedSuggestedSwitch}
-                    alignmentEndPoint={selectedSuggestedSwitch.alignmentEndPoint}
-                    onSuggestedSwitchChange={onSuggestedSwitchChange}
-                />
-            )}
+            {selectedSuggestedSwitch?.alignmentEndPoint &&
+                switchStructure &&
+                onSuggestedSwitchChange && (
+                    <GeometrySwitchLinkingSuggestedInfobox
+                        contentVisible={visibilities.suggestedSwitch}
+                        onContentVisibilityChange={() =>
+                            onVisibilityChange({
+                                ...visibilities,
+                                suggestedSwitch: !visibilities.suggestedSwitch,
+                            })
+                        }
+                        suggestedSwitch={selectedSuggestedSwitch}
+                        suggestedSwitchStructure={switchStructure}
+                        alignmentEndPoint={selectedSuggestedSwitch.alignmentEndPoint}
+                        onSuggestedSwitchChange={onSuggestedSwitchChange}
+                    />
+                )}
             <Infobox
                 contentVisible={visibilities.linking}
                 onContentVisibilityChange={() =>
@@ -237,9 +224,10 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
                                     suggestedSwitch={suggestedSwitch}
                                     onShowAddSwitchDialog={() => setShowAddSwitchDialog(true)}
                                 />
-                                {suggestedSwitch && (
+                                {suggestedSwitch && switchStructure && (
                                     <SwitchJointInfoboxContainer
                                         suggestedSwitch={suggestedSwitch}
+                                        suggestedSwitchStructure={switchStructure}
                                         publishType={publishType}
                                     />
                                 )}
@@ -250,10 +238,10 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
                                     {t('tool-panel.switch.geometry.zoom-closer')}
                                 </div>
                             )}
-                            {suggestedSwitch && (
+                            {suggestedSwitch && switchStructure && (
                                 <GeometrySwitchLinkingErrors
                                     selectedLayoutSwitchStructure={selectedLayoutSwitchStructure}
-                                    suggestedSwitch={suggestedSwitch}
+                                    suggestedSwitchStructure={switchStructure}
                                     switchTypeMatch={switchTypeMatch}
                                     onConfirmChanged={(confirmed) =>
                                         setSwitchTypeDifferenceIsConfirmed(confirmed)
@@ -286,7 +274,7 @@ const GeometrySwitchLinkingInfobox: React.FC<GeometrySwitchLinkingInfoboxProps> 
                 <SwitchEditDialogContainer
                     onClose={() => setShowAddSwitchDialog(false)}
                     onSave={handleSwitchSave}
-                    prefilledSwitchStructureId={selectedSuggestedSwitch.switchStructure.id}
+                    prefilledSwitchStructureId={selectedSuggestedSwitch.switchStructureId}
                 />
             )}
         </React.Fragment>
