+++ ui/src/tool-panel/switch/switch-joint-infobox-container.tsx
@@ -1,28 +1,33 @@
 import * as React from 'react';
 import SwitchJointInfobox from 'tool-panel/switch/switch-joint-infobox';
-import { asTrackLayoutSwitchJointConnection } from 'linking/linking-utils';
+import {
+    suggestedSwitchJointsAsLayoutSwitchJointConnections,
+    suggestedSwitchTopoLinksAsTopologicalJointConnections,
+} from 'linking/linking-utils';
 import { SuggestedSwitch } from 'linking/linking-model';
-import { PublishType } from 'common/common-model';
+import { PublishType, SwitchStructure } from 'common/common-model';
 
 type SwitchJointInfoboxContainerProps = {
     suggestedSwitch: SuggestedSwitch;
+    suggestedSwitchStructure: SwitchStructure;
     publishType: PublishType;
 };
 
 export const SwitchJointInfoboxContainer: React.FC<SwitchJointInfoboxContainerProps> = ({
     suggestedSwitch,
+    suggestedSwitchStructure,
     publishType,
 }) => {
-    const jointConnections = suggestedSwitch
-        ? suggestedSwitch.joints.map((joint) => asTrackLayoutSwitchJointConnection(joint))
-        : undefined;
-    const switchAlignments = suggestedSwitch.switchStructure.alignments;
+    const jointConnections = suggestedSwitchJointsAsLayoutSwitchJointConnections(suggestedSwitch);
+    const switchAlignments = suggestedSwitchStructure.alignments;
+    const topologicalJointConnections =
+        suggestedSwitchTopoLinksAsTopologicalJointConnections(suggestedSwitch);
 
     return jointConnections ? (
         <SwitchJointInfobox
             switchAlignments={switchAlignments}
             jointConnections={jointConnections}
-            topologicalJointConnections={suggestedSwitch?.topologicalJointConnections}
+            topologicalJointConnections={topologicalJointConnections}
             publishType={publishType}
         />
     ) : (
